# packages/beat/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Beat Detective Field Kit API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/cases/{caseId}/files:
    parameters:
      - name: caseId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: listCaseFiles
      tags: [files]
      summary: List files attached to a case
      parameters:
        - name: sourceType
          in: query
          required: false
          schema:
            type: string
            enum: [photo, audio, note, scan, upload, drive]
      responses:
        "200":
          description: File list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseFileListResponse"
    post:
      operationId: uploadCaseFile
      tags: [files]
      summary: Upload a file to a case (multipart)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                sourceType:
                  type: string
                  enum: [photo, audio, note, scan, upload, drive]
                caption:
                  type: string
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseFile"
  /v1/cases/{caseId}/files/{fileId}:
    parameters:
      - name: caseId
        in: path
        required: true
        schema:
          type: string
      - name: fileId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getCaseFile
      tags: [files]
      summary: Get metadata for a case file
      responses:
        "200":
          description: File metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseFile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCaseFile
      tags: [files]
      summary: Delete a case file
      responses:
        "204":
          description: Deleted
  /v1/cases/{caseId}/files/{fileId}/content:
    parameters:
      - name: caseId
        in: path
        required: true
        schema:
          type: string
      - name: fileId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getCaseFileContent
      tags: [files]
      summary: Serve the raw bytes of a case file
      responses:
        "200":
          description: File bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Run lifecycle ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string

    # ── User ─────────────────────────────────────────────────────────────────
...
    CaseFile:
      type: object
      required: [id, caseId, filename, mimeType, sizeBytes, storageUrl, createdAt]
      properties:
        id:
          type: string
        caseId:
          type: string
        sourceType:
          type: string
          enum: [photo, audio, note, scan, upload, drive]
          nullable: true
        filename:
          type: string
        originalName:
          type: string
          nullable: true
        mimeType:
          type: string
        sha256:
          type: string
          nullable: true
        sizeBytes:
          type: integer
        storageUrl:
          type: string
        exif:
          type: object
          additionalProperties: true
          nullable: true
        gps:
          type: object
          additionalProperties: true
          nullable: true
        capturedAt:
          type: string
          format: date-time
          nullable: true
        detectedLanguage:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
        transcript:
          type: string
          nullable: true
        autoTagJson:
          type: object
          additionalProperties: true
          nullable: true
        createdAt:
          type: string
          format: date-time
    CaseFileListResponse:
      type: object
      required: [files]
      properties:
        files:
          type: array
          items:
            $ref: "#/components/schemas/CaseFile"

    # ── Run ──────────────────────────────────────────────────────────────────
...